+++ apps/docs/src/app/core/component-docs/busy-indicator/examples/busy-indicator-wrapper-example.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-busy-indicator-wrapper-example',
+    templateUrl: './busy-indicator-wrapper-example.component.html'
+})
+export class BusyIndicatorWrapperExampleComponent {
+    loading = true;
+}