+++ apps/docs/src/app/core/component-docs/select/examples/select-view-value-example/select-view-value-example.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-select-view-value-example',
-    templateUrl: './select-view-value-example.component.html'
-})
-export class SelectViewValueExampleComponent {
-    selectOneValue: string;
-}