+++ apps/docs/src/app/core/component-docs/select/select-header/select-header.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'fd-select-header',
-    templateUrl: './select-header.component.html',
-    styleUrls: ['./select-header.component.scss']
+    templateUrl: './select-header.component.html'
 })
 export class SelectHeaderComponent {}