+++ libs/core/src/lib/popover/popover-directive/popover.directive.ts
@@ -281,6 +281,7 @@ export class PopoverDirective implements OnInit, OnDestroy, OnChanges {
         this.appRef.attachView(this.containerRef.hostView);
         const setupRef = this.containerRef.instance.isSetup.subscribe(() => {
             this.createPopper();
+            this.updatePopper();
             setupRef.unsubscribe();
         });
 
@@ -304,12 +305,10 @@ export class PopoverDirective implements OnInit, OnDestroy, OnChanges {
 
     private addTriggerListeners(): void {
         if (this.triggers && this.triggers.length > 0) {
-            this.triggers.forEach((trigger) => {
-                this.eventRef.push(
-                    this.renderer.listen(this.elRef.nativeElement, trigger, () => {
-                        this.toggle();
-                    })
-                );
+            this.triggers.forEach(trigger => {
+                this.eventRef.push(this.renderer.listen(this.elRef.nativeElement, trigger, () => {
+                    this.toggle();
+                }));
             });
         }
     }