+++ libs/platform/src/lib/components/action-button-group/action-button-group.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ActionButtonGroupComponent } from './action-button-group.component';
+import { ActionBarModule } from '@fundamental-ngx/core';
+@NgModule({
+    declarations: [ActionButtonGroupComponent],
+    imports: [CommonModule, ActionBarModule],
+    exports: [ActionButtonGroupComponent]
+})
+export class PlatformActionButtonGroupModule { }