+++ libs/platform/src/lib/components/action-button-group/public_api.ts
@@ -1,0 +1,2 @@
+export * from './action-button-group.module';
+export * from './action-button-group.component';