+++ libs/platform/src/lib/platform.module.ts
@@ -3,7 +3,8 @@ import { CommonModule } from '@angular/common';
 import { FundamentalNgxCoreModule } from '@fundamental-ngx/core';
 import { PlatformButtonModule } from './components/button/button.module';
 import { PlatformSearchFieldModule } from './components/search-field/search-field.module';
-import { ActionBarModule } from './components/action-bar/action-bar.module';
+import { PlatformActionBarModule } from './components/action-bar/action-bar.module';
+import { PlatformActionButtonGroupModule } from './components/action-button-group/action-button-group.module';
 import { PlatformMenuModule } from './components/menu/menu.module';
 import { PlatformSelectModule } from './components/select/select.module';
 import { PlatformLinkModule } from './components/link/link.module';
@@ -16,13 +17,14 @@ import { PlatformSplitMenuButtonModule } from './components/split-menu-button/sp
     exports: [
         PlatformButtonModule,
         PlatformSearchFieldModule,
-        ActionBarModule,
+        PlatformActionBarModule,
+        PlatformActionButtonGroupModule,
         PlatformMenuModule,
         PlatformSelectModule,
         PlatformLinkModule,
         PlatformRadioGroupModule,
         PlatformMenuButtonModule,
-        PlatformSplitMenuButtonModule,
-    ],
+        PlatformSplitMenuButtonModule
+    ]
 })
-export class FundamentalNgxPlatformModule {}
+export class FundamentalNgxPlatformModule { }